+++ src/main/java/chikachi/discord/config/message/Patterns.java
@@ -6,6 +6,7 @@ class Patterns {
     static final Pattern everyonePattern = Pattern.compile("(^|\\W)@everyone\\b");
     static final Pattern boldPattern = Pattern.compile("\\*\\*(.*)\\*\\*");
     static final Pattern italicPattern = Pattern.compile("\\*(.*)\\*");
+    static final Pattern italicMePattern = Pattern.compile("\\*(.*)\\*");
     static final Pattern underlinePattern = Pattern.compile("__(.*)__");
     static final Pattern lineThroughPattern = Pattern.compile("~~(.*)~~");
     static final Pattern singleCodePattern = Pattern.compile("`(.*)`");